move two github functions in the github module
